Instagram/requirements.txt
fastapi
uvicorn
snowflake-sqlalchemy
pymysql
sqlalchemy